학생관리에서 등록 수정 삭제 상세보기 구현을 해놓았고 조회할때 조회목록을 센터에 둘 예정
